--- a/foundation.cabal
+++ b/foundation.cabal
@@ -131,13 +131,13 @@ Test-Suite test-foundation
   type:              exitcode-stdio-1.0
   hs-source-dirs:    tests
   Main-is:           Tests.hs
-  Other-modules:     Encoding
-                     Parser
-                     Test.Utils.Foreign
+  Other-modules:     Test.Utils.Foreign
                      Test.Data.List
                      Test.Data.Unicode
                      Test.Foundation.Collection
                      Test.Foundation.Number
+                     Test.Foundation.Encoding
+                     Test.Foundation.Parser
   Build-Depends:     base >= 3 && < 5
                    , mtl
                    , QuickCheck