… tests/Test/Utils/Foreign.hs
… tests/Tests.hs