Testing has been a bit long so far:

Before:

```stack test  128.66s user 5.99s system 94% cpu 2:21.96 total```

After:

```stack test  85.86s user 5.30s system 97% cpu 1:33.84 total```

Also, contributors will find that, when working on a specific test, it will be a bit faster as well to generate the test.

I have also added a couple of test of the `Number` type class, mainly to validate the property specified in the documentation (associativity of the `Additive`...)