@NicolasDP 2016-09-03
The added tests are:

      Additive
        a + azero == a:                                       OK
          +++ OK, passed 100 tests.
        azero + a == a:                                       OK
          +++ OK, passed 100 tests.
        a + b == b + a:                                       OK
          +++ OK, passed 100 tests.
      Multiplicative
        a * midentity == a:                                   OK
          +++ OK, passed 100 tests.
        midentity * a == a:                                   OK
          +++ OK, passed 100 tests.
      Divisible
        (x `div` y) * y + (x `mod` y) == x:                   OK
          +++ OK, passed 100 tests.