@adinapoli 2016-09-03
Just throwing some brain dump here: what would also be useful is a QC property which would ensure that `forall arbitrary $ \(lst :: [ty) -> lst == toList (fromList lst)` 😉