@NicolasDP 2016-09-03
@adinapoli you mean like this one:

```
  Bitmap
    equality
      x == x:                                                 OK
        +++ OK, passed 100 tests.
      x == y:                                                 OK
        +++ OK, passed 100 tests.
    ordering
      x `compare` y:                                          OK (0.01s)
        +++ OK, passed 100 tests.
    IsList
      fromList . toList == id:                                OK
        +++ OK, passed 100 tests.
```

I have also re-oder few tests