@adinapoli 2016-09-04
Yes, exactly like that! (Sorry, I missed that property while reading the
source) ;)

On 3 Sep 2016 11:20 p.m., "Nicolas DI PRIMA" <notifications@github.com>
wrote:

> Merged #107 <https://github.com/haskell-foundation/foundation/pull/107>.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/haskell-foundation/foundation/pull/107#event-777464324>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAa-s8SvUDY-2kHhb3oS-4mLxLs5MfiYks5qmeS4gaJpZM4J0TqK>
> .
>